wip: keymat_v2: Cache initial IKE messages for auth octets
This avoids pre-generating the message to be sent and supports fragments
as used for IKE_AUX.

In scenarios with IKE_AUX this basically changes the auth octets as follows:

  InitiatorSignedOctets = RealMessage1(INIT) | RealMessage3(AUX) | ...
                          NonceRData | MACedIDForI

and

  ResponderSignedOctets = RealMessage2(INIT) | RealMessage4(AUX) | ...
                          NonceIData | MACedIDForR

wip: TKM doesn't work yet. It might be better to store these packets on the
ike_sa (for increased flexiblity, e.g. with regards to conftest, we
currently don't use them directly in get_auth_octets anyway), or the
task manager (however, it might be tricky to get them to the ike-auth
task from there).
wip: Since this requires keeping around quite some data, alternatives would
be to hash the message (with some negotiated or fixed hash function) or
applying the PRF (if it is QC-safe, e.g. with a zero key or the ones we
derived from DH).